#10 criando a seed para usuarios de teste

+++ src/database/migrations/20240325014454-create-table-usuario.js
@@ -25,7 +25,7 @@ module.exports = {
             },
             celular: {
                 allowNull: false,
-                type: types.INTEGER
+                type: types.BIGINT
             },
             created_at: {
                 allowNull: false,
